🚑️ Fix everything needs to update, logic was inverted
