+++ src/cache.rs
@@ -10,7 +10,7 @@ use rkyv::{Archive, Deserialize, Serialize, deserialize, rancor};
 use tracing::{debug, instrument};
 use xxhash_rust::xxh64::Xxh64;
 
-use crate::{config::AppConfig, log::anonymize};
+use crate::{config::AppConfig, core::Checksums, log::anonymize};
 
 #[derive(thiserror::Error, Debug)]
 pub enum CacheError {
@@ -23,8 +23,17 @@ pub enum CacheError {
 /// Represents database of file cache.
 #[derive(Archive, Deserialize, Serialize, Debug, Default)]
 #[rkyv(compare(PartialEq), derive(Debug))]
-pub struct FileCacheDB {
+pub struct FileCacheDb {
     pub entries: BTreeMap<u64, CacheEntry>,
+}
+
+impl FileCacheDb {
+    pub fn is_cache_valid(&self, inode: &u64, checksums: &Checksums) -> bool {
+        self.entries
+            .get(inode)
+            .map(|entry| checksums.contains(entry.hash()))
+            .unwrap_or(false)
+    }
 }
 
 /// Snapshot of the file when it was last hashed.
@@ -58,7 +67,7 @@ impl CacheEntry {
 
 /// Gets up-to-date file cache.
 #[instrument(skip(config), fields(path = %anonymize(config.cache_db_path())))]
-pub fn sync(config: &AppConfig) -> Result<BTreeMap<u64, CacheEntry>, CacheError> {
+pub fn sync(config: &AppConfig) -> Result<FileCacheDb, CacheError> {
     // Load existing cache database
     let mut cache = load_cache_db(config.cache_db_path()).unwrap_or_default();
 
@@ -66,11 +75,11 @@ pub fn sync(config: &AppConfig) -> Result<BTreeMap<u64, CacheEntry>, CacheError>
         save_cache_db(&cache, config.cache_db_path())?;
     }
 
-    Ok(cache.entries)
+    Ok(cache)
 }
 
 /// Updates cache entries based on current filesystem state.
-fn update_cache(cache: &mut FileCacheDB, mods_dir: &Path) -> io::Result<bool> {
+fn update_cache(cache: &mut FileCacheDb, mods_dir: &Path) -> io::Result<bool> {
     let mut current_keys = HashSet::new();
     let mut updated = false;
 
@@ -130,15 +139,15 @@ fn should_rehash(entries: &BTreeMap<u64, CacheEntry>, key: &u64, mtime: i64, siz
 }
 
 /// Loads cache database from disk using rkyv.
-fn load_cache_db(cache_path: &Path) -> Result<FileCacheDB, CacheError> {
+fn load_cache_db(cache_path: &Path) -> Result<FileCacheDb, CacheError> {
     let bytes = fs::read(cache_path)?;
-    let archived = rkyv::access::<ArchivedFileCacheDB, rancor::Error>(&bytes)?;
-    let cache = deserialize::<FileCacheDB, rancor::Error>(archived)?;
+    let archived = rkyv::access::<ArchivedFileCacheDb, rancor::Error>(&bytes)?;
+    let cache = deserialize::<FileCacheDb, rancor::Error>(archived)?;
     Ok(cache)
 }
 
 /// Saves cache database to disk using rkyv.
-fn save_cache_db(cache: &FileCacheDB, cache_path: &Path) -> Result<(), CacheError> {
+fn save_cache_db(cache: &FileCacheDb, cache_path: &Path) -> Result<(), CacheError> {
     let bytes = rkyv::to_bytes::<rancor::Error>(cache)?;
     let mut file = fs::OpenOptions::new()
         .create(true)
