+++ src/core.rs
@@ -1,3 +1,4 @@
+pub mod checksum;
 pub mod everest;
 pub mod loader;
 pub mod local;
@@ -7,5 +8,8 @@ pub mod registry;
 pub mod resolver;
 pub mod update;
 
+pub use checksum::{
+    Checksum, ChecksumError, ChecksumVerificationError, ChecksumVerifier, Checksums,
+};
 pub use local::LocalMod;
 pub use mod_file::ModFile;
